refactor: remove `__tests__` folder
