+++ tsconfig.json
@@ -19,5 +19,5 @@
     }
   },
   "include": ["src/**/*"],
-  "exclude": ["src/**/__tests__/*"]
+  "exclude": ["src/**/*.spec.ts"]
 }
